net: Moved Ethernet specific addresses from MAC to Eth
